Nest all code into a common parquet/ top level source directory
