+++ src/parquet/encodings/delta-bit-pack-encoding.h
@@ -15,7 +15,7 @@
 #ifndef PARQUET_DELTA_BIT_PACK_ENCODING_H
 #define PARQUET_DELTA_BIT_PACK_ENCODING_H
 
-#include "encodings.h"
+#include "parquet/encodings/encodings.h"
 
 namespace parquet_cpp {
 
@@ -30,7 +30,7 @@ class DeltaBitPackDecoder : public Decoder {
 
   virtual void SetData(int num_values, const uint8_t* data, int len) {
     num_values_ = num_values;
-    decoder_ = impala::BitReader(data, len);
+    decoder_ = BitReader(data, len);
     values_current_block_ = 0;
     values_current_mini_block_ = 0;
   }
@@ -94,7 +94,7 @@ class DeltaBitPackDecoder : public Decoder {
     return max_values;
   }
 
-  impala::BitReader decoder_;
+  BitReader decoder_;
   uint64_t values_current_block_;
   uint64_t num_mini_blocks_;
   uint64_t values_per_mini_block_;
@@ -111,4 +111,3 @@ class DeltaBitPackDecoder : public Decoder {
 }
 
 #endif
-
