+++ src/parquet/encodings/delta-byte-array-encoding.h
@@ -15,7 +15,7 @@
 #ifndef PARQUET_DELTA_BYTE_ARRAY_ENCODING_H
 #define PARQUET_DELTA_BYTE_ARRAY_ENCODING_H
 
-#include "encodings.h"
+#include "parquet/encodings/encodings.h"
 
 namespace parquet_cpp {
 
@@ -70,4 +70,3 @@ class DeltaByteArrayDecoder : public Decoder {
 }
 
 #endif
-
