+++ src/parquet/encodings/delta-length-byte-array-encoding.h
@@ -15,7 +15,7 @@
 #ifndef PARQUET_DELTA_LENGTH_BYTE_ARRAY_ENCODING_H
 #define PARQUET_DELTA_LENGTH_BYTE_ARRAY_ENCODING_H
 
-#include "encodings.h"
+#include "parquet/encodings/encodings.h"
 
 namespace parquet_cpp {
 
@@ -59,4 +59,3 @@ class DeltaLengthByteArrayDecoder : public Decoder {
 }
 
 #endif
-
