+++ src/parquet/encodings/plain-encoding.h
@@ -15,7 +15,7 @@
 #ifndef PARQUET_PLAIN_ENCODING_H
 #define PARQUET_PLAIN_ENCODING_H
 
-#include "encodings.h"
+#include "parquet/encodings/encodings.h"
 
 namespace parquet_cpp {
 
@@ -79,4 +79,3 @@ class PlainDecoder : public Decoder {
 }
 
 #endif
-
